chore: move old endpoints to a new file
